Declare the app guide's deviations where both sides can execute them
app-template mirrors `docs/agents/building-apps.md` into its AGENTS.md, and
that mirror is about to stop being hand-kept: it is generated there now. Both
programs need the same three substitutions — this gate applies them to compare,
that generator applies them to write — and two transcriptions of one list would
move the drift a level down, into the least visible place there is. A generator
and a gate that disagree produce a file that fails a check nobody can fix from
either side alone.

So `GUIDE_DEVIATIONS` moves out of this script into
`.github/shared/app-guide-deviations.mjs`, together with `guideBody` and the
apply function, and the file joins the SHARED list with app-template as its
consumer. This gate imports what it used to declare; nothing it checks changes.
